+++ onyx/crates/ember-core/examples/hello_tcp.rs
@@ -1,10 +1,18 @@
 use ember_core::server::Server;
 use ember_http::request::Request;
-
+use ember_core::json::Json;
 
 use ember_core::{extractor::FromRequest, path::Path,
 query::Query,
 header::Header,};
+
+use serde::Serialize;
+
+#[derive(Serialize)]
+struct User {
+    id: u32,
+    name: String,
+}
 
 fn home(_: Request) -> &'static str {
     "Hello from Onyx!"
@@ -32,6 +40,14 @@ fn host(request: Request) -> String {
     format!("Host: {}", host)
 }
 
+
+fn api(_: Request) -> Json<User> {
+    Json(User {
+        id: 1,
+        name: "Soham".into(),
+    })
+}
+
 fn main() {
     let mut server = Server::new("127.0.0.1:8080");
 
@@ -40,6 +56,7 @@ fn main() {
     server.router_mut().get("/users/:id", user);
     server.router_mut().get("/search", search);
     server.router_mut().get("/host", host);
+    server.router_mut().get("/api/user", api);
     if let Err(error) = server.run() {
         println!("Server error: {:?}", error);
     }